derp

+++ A3/ext2_cp.c
@@ -274,7 +274,7 @@ int main(int argc, char **argv) {
     for (i = 0; i < 16; i+=1, bbmap +=1) {
         for (pos = 0; pos < 8; pos+=1) {
 			if (block_bitmap[(8 * i) + pos] == 1) {
-                *bbmap |= pow(2,pos);
+                *bbmap |= (int) pow(2,pos);
 			}
         }
     }
